BE - move login to components, add index redirection to login
