+++ cms/index.php
@@ -1,0 +1,1 @@
+<?php header('Location: login.php'); ?>
